make `Test` use a new `TestHost`
