+++ stdlib/Cargo.toml
@@ -32,6 +32,7 @@ with-debug-info = []
 
 [dependencies]
 assembly = { package = "miden-assembly", path = "../assembly", version = "0.13", default-features = false }
+vm-core = { package = "miden-core", path = "../core", version = "0.13", default-features = false }
 
 [dev-dependencies]
 blake3 = "1.5"
@@ -48,7 +49,6 @@ serde_json = "1.0"
 sha2 = "0.10"
 sha3 = "0.10"
 test-utils = { package = "miden-test-utils", path = "../test-utils" }
-vm-core = { package = "miden-core", path = "../core", version = "0.13", default-features = false }
 winter-air = { package = "winter-air", version = "0.12" }
 winter-fri = { package = "winter-fri", version = "0.12" }
